requirements-dev.txt
black==23.3.0
factory-boy==3.2.1
pip-tools
## pygraphviz==1.11
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-socket==0.6.0
pytest-xdist==3.3.1

